+++ extensions/how-long-to-beat/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Howlongtobeat Changelog
 
+## [Fixes] - 2026-02-24
+
+- Fix: update to new API endpoint
+
 ## [Fixes] - 2026-01-22
 
 - Fix: search error when fetching games by search